rename to avoid conflict
The conflict only appears when testing individual modules in ipython like this:

    %run nsls2ptycho/core/CSX_databroker.py

It does not affect normal usage. However, it's still good to have a different
module name to make code search a bit easier.
